Gilbert Ruby V0.6.7

+++ game_state/level1.rb
@@ -1,33 +1,75 @@
 class Level1 < Chingu::GameState
+  attr_reader :music, :sonido
   traits :viewport, :timer
   GAME_OBJECTS = [Base, Moneda, Plataforma, Plataforma2]
-  def initialize
+  def initialize(options = {})
     super
-    @player_name = previous_game_state().player_name.capitalize
+    @player_name = options[:player_name].capitalize
+    @sonido = options[:sonido]
 
-    #self.input = {e: :edit}
-    @background = Image['paisaje.png']
+    # self.input = {e: :edit}
+    @background = Image['Level1Bg.png']
     self.viewport.lag = 0
     self.viewport.game_area = [0, 0, 3000, 1000]
-    load_game_objects
-    @sonido = Song.new("media/Songs/SoundTrack.mp3")
-    @loser_song = Song.new("media/Songs/gameover.ogg")
-    play_song
-    @gilbert = Gilbert.create(x: 50, y: 900)
-    @contador = 0
-    @marcador = Chingu::Text.new("#{@player_name}:  Puntaje: #{@contador}", x: 5, y: 5, size: 35, color: Gosu::Color::BLACK)
+
+    @gilbert = Gilbert.create(x: 50, y: 900, limited: self.viewport.game_area.width)
+    setup
     @plataformas = Plataforma2.all
+
+    @music = Song.new("media/Songs/Music Level1.mp3")
+    @music.play(true) if @sonido
+    @loser_song = Song.new("media/Songs/gameover.ogg")
+
+    randon = ramdon_location
+    @boton = Boton.create(x: randon[0], y: randon[1] + 18)
+    @plataforma_meta = Meta.create(x: randon[0], y: randon[1] + 45)
+
+    enable_meta(false)
+
+    @marco_name = Image['MarcoName.png']
+    @marco_score = Image['MarcoPoint.png']
+
+    @score = 0
+    @msjName = Chingu::Text.new("#{@player_name}", x: 85, y: 25, size: 30, color: Gosu::Color::WHITE)
+    @msjScore = Chingu::Text.new("#{@score}", x: $window.width - 130, y: 30, size: 35)
+    @mensaje = Chingu::Text.new("Has encontrado todas las monedas", x: 320, y: 20, size: 25, color: Gosu::Color::GREEN)
+    @mensaje2 = Chingu::Text.new("Encuentra el Boton", x: @msjName.x, y: 45, size: 30, color: Gosu::Color::YELLOW)
   end
 
   def lose
-  	play_song(nil)
-    @loser_song.play
+    meta_random_location
+  	@music.pause if @sonido
+    @loser_song.play if @sonido
     @gilbert.x = 50
     @gilbert.y = 900
     GAME_OBJECTS.each(&:destroy_all)
-    load_game_objects
-    push_game_state(Puntajes.new(player: @player_name, score: @contador))
-    @contador = 0
+    load_game_objects(file: @level)
+    push_game_state(Puntajes.new(player_name: @player_name, score: @score))
+    @score = 0
+  end
+
+  def enable_meta(boolean = true)
+    if boolean
+      @boton.show!
+      @plataforma_meta.show!
+    else
+      @boton.hide!
+      @plataforma_meta.hide!
+    end
+  end
+
+  def meta_random_location
+    randon = ramdon_location
+    @boton.x, @boton.y = randon[0], randon[1] + 18
+    @plataforma_meta.x, @plataforma_meta.y = randon[0], randon[1] + 45
+  end
+
+  def win
+    meta_random_location
+    GAME_OBJECTS.each(&:destroy_all)
+    @music.pause
+  	push_game_state(Level2.new(player_name: @player_name, score: @score, sonido: @sonido))
+    @score = 0
   end
 
   def ramdon_location
@@ -41,42 +83,64 @@ class Level1 < Chingu::GameState
   	Moneda.create(x: array[0], y: array[1])
   end
 
-  def play_song(play = true)
-    case play
-    when true
-      @sonido.play(true)
-    when false
-      @sonido.pause
-    when nil
-      @sonido.stop
-    end
+  def setup
+    super
+    @level = File.join(ROOT, "levels", self.filename + ".yml")
+    load_game_objects(file: @level)
+    
+    @gilbert.swicht = true
+    
+    @gilbert.x = 50
+    @gilbert.y = 900
   end
 
   # def edit
-  #   push_game_state(Chingu::GameStates::Edit.new(except: [Gilbert, PulsatingText]))
+  #   push_game_state(Chingu::GameStates::Edit.new(except: [Gilbert, PulsatingText, Boton, BaseLv2, PlataformaMLv2, PlataformaSLv2]))
   # end
 
   def draw
-    super
     @background.draw(0, 0, 0)
-    @marcador.draw()
+    @marco_name.draw(0, 0, 0)
+    @marco_score.draw($window.width - 170, 0, 0)
+    super
+    @msjName.draw()
+    @msjScore.draw()
+    if @score == 15
+      Plataforma2.all.select{|e| e.x == @plataforma_meta.x && e.y == @plataforma_meta.y}[0].hide!
+      enable_meta
+      @msjName.y = 15
+      @mensaje.draw
+      @mensaje2.draw
+    elsif @score < 15
+      Plataforma2.all.select{|e| e.x == @plataforma_meta.x && e.y == @plataforma_meta.y}[0].show!
+      enable_meta(false)
+    end
   end
 
   def update
     super
-    $window.caption = "Gilbert Ruby\tFPS:#{self.fps}"
-    @marcador.text = "#{@player_name}:  Puntaje: #{@contador}"
+    $window.caption = "Gilbert Ruby\tFPS:#{self.fps}\t Objects: #{current_game_state.game_objects.size}"
+    @msjScore.text = "#{@score}"
     self.viewport.center_around(@gilbert)
-    @gilbert.each_collision(Moneda) do |gilbert, moneda|
-      moneda.play_sound
-      @contador += 1
-      moneda.destroy
-      create_coin(ramdon_location)
+
+    if @score < 15
+    	@gilbert.each_collision(Moneda) do |gilbert, moneda|
+	      moneda.play_sound
+	      @score += 1
+	      moneda.destroy
+	      create_coin(ramdon_location)
+    	end
+    end
+
+    if @score == 15
+    	Moneda.destroy_all
+    	@gilbert.each_collision(Boton) do |gilbert, boton|
+    		win
+    	end
     end
 
     if self.viewport.outside_game_area?(@gilbert)
       if @gilbert.y >= self.viewport.game_area.height
-        @gilbert.y = self.viewport.game_area.height + 100
         lose
       end
     end
